+++ packages/core/form/src/Form.js
@@ -73,7 +73,9 @@ class Form extends React.Component<Props, State> {
 
   formRef = React.createRef();
 
-  form = createFinalForm(this.props.onSubmit, this.formRef);
+  onSubmitProxy = (...args: Array<any>) => this.props.onSubmit(...args);
+
+  form = createFinalForm(this.onSubmitProxy, this.formRef);
 
   state = {
     dirty: false,