+++ packages/help/help/examples/4-Help-Right-Side-Panel-Algolia-.tsx
@@ -10,7 +10,11 @@ import {
 } from '@atlaskit/right-side-panel';
 
 import LocaleIntlProvider from '../example-helpers/LocaleIntlProvider';
-import { ButtonsWrapper, FooterContent } from './utils/styled';
+import {
+  ButtonsWrapper,
+  FooterContent,
+  ExampleDefaultContent,
+} from './utils/styled';
 
 import Help from '../src';
 
@@ -140,7 +144,9 @@ export default class extends React.Component {
                       </FooterContent>
                     }
                   >
-                    <span>Default content</span>
+                    <ExampleDefaultContent>
+                      <span>Default content</span>
+                    </ExampleDefaultContent>
                   </Help>
                 </LocaleIntlProvider>
               </RightSidePanel>