+++ packages/help/help/src/components/HelpContent.tsx
@@ -5,7 +5,13 @@ import Header from './Header';
 import Search from './Search';
 import ArticleComponent from './Article';
 
-import { Container, HelpBody, HelpFooter, Section } from './styled';
+import {
+  Container,
+  HelpBody,
+  HelpFooter,
+  Section,
+  DefaultContent,
+} from './styled';
 
 export interface Props {}
 
@@ -20,7 +26,9 @@ export const HelpContent = (props: Props & HelpContextInterface) => {
           <HelpBody>
             {help.isSearchVisible() && <Search />}
             <ArticleComponent />
-            {help.defaultContent}
+            <DefaultContent isArticleVisible={help.isArticleVisible()}>
+              {help.defaultContent}
+            </DefaultContent>
           </HelpBody>
           {help.isFooter() ? <HelpFooter>{help.footer}</HelpFooter> : null}
         </Section>