+++ website/src/containers/App.tsx
@@ -1,18 +1,18 @@
-import * as React from 'react';
+import React from 'react';
 import Media from 'react-media';
 import GlobalTheme from '@atlaskit/theme';
 import Page from '@atlaskit/page';
 import { RouteProps } from 'react-router';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import Nav from './Nav';
-import MobileNav from './MobileNav';
 import { DESKTOP_BREAKPOINT_MIN } from '../constants';
 import FullscreenExamples from '../pages/Examples';
 import { modalRoutes, pageRoutes } from '../routes';
 import ScrollHandler from '../components/ScrollToTop';
 import ErrorBoundary from '../components/ErrorBoundary';
 import AnalyticsListeners from '../components/Analytics/AnalyticsListeners';
+import DesktopNav from './DesktopNav';
+import MobileNav from './MobileNav';
 
 export default () => {
   return (
@@ -31,7 +31,7 @@ export default () => {
                   render={appRouteDetails => (
                     <Page
                       navigation={
-                        isDesktop ? <Nav {...appRouteDetails} /> : false
+                        isDesktop ? <DesktopNav {...appRouteDetails} /> : false
                       }
                     >
                       {!isDesktop && (