+++ packages/help/help/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## 0.1.0
 
+## 0.17.4
+
+### Patch Changes
+
+- [patch][22d5dbc2bb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/22d5dbc2bb):
+
+  fixed CSS issue of the ArticleListItem icons
+
+## 0.17.3
+
+### Patch Changes
+
+- [patch][c0c49e4c2b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c0c49e4c2b):
+
+  CSS fixes + added link icon to ArticleListItem
+
 ## 0.17.2
 
 ### Patch Changes